Rename controller → controller_impl
